Move flows/tests to flows/routers/tests
