+++ cmd/docgen/main_test.go
@@ -13,7 +13,7 @@ func TestDocGeneration(t *testing.T) {
 	// tests run from the same working directory as the test file, so two directories up is our goflow root
 	path := "../../"
 	buildDocSet(path, "excellent", "@function", handleFunctionDoc, session)
-	buildDocSet(path, "flows/tests", "@test", handleFunctionDoc, session)
+	buildDocSet(path, "flows/routers/tests", "@test", handleFunctionDoc, session)
 	buildDocSet(path, "flows/actions", "@action", handleActionDoc, session)
 	buildDocSet(path, "flows/events", "@event", handleEventDoc, session)
 }
